[TRL-175] chore: 여행 관련 조회 컨트롤러 패키지를 레이어드 아키텍처에 맞게 이동
