+++ src/main/java/com/cosain/trilo/trip/query/presentation/trip/TripTemporaryStorageQueryController.java
@@ -1,4 +1,4 @@
-package com.cosain.trilo.trip.query.adapter.in.api.trip;
+package com.cosain.trilo.trip.query.presentation.trip;
 
 import com.cosain.trilo.common.exception.NotImplementedException;
 import lombok.extern.slf4j.Slf4j;
