+++ src/test/java/com/cosain/trilo/unit/trip/query/presentation/schedule/SingleScheduleQueryControllerTest.java
@@ -1,5 +1,4 @@
-package com.cosain.trilo.unit.trip.query.adapter.in.api.trip;
-
+package com.cosain.trilo.unit.trip.query.presentation.schedule;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -20,8 +19,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
-@DisplayName("특정 사용자 여행 목록 API 테스트")
-class TripperTripListQueryControllerTest {
+@DisplayName("일정 단건 조회 API 테스트")
+class SingleScheduleQueryControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
@@ -36,8 +35,8 @@ class TripperTripListQueryControllerTest {
     @Test
     @DisplayName("인증된 사용자 요청 -> 미구현 500")
     @WithMockUser
-    public void findTripperTripList_with_authorizedUser() throws Exception {
-        mockMvc.perform(get("/api/trips?tripper-id=1"))
+    public void findSingleSchedule_with_authorizedUser() throws Exception {
+        mockMvc.perform(get("/api/schedule-places/1"))
                 .andDo(print())
                 .andExpect(status().isInternalServerError())
                 .andExpect(jsonPath("$.errorCode").exists())
@@ -47,8 +46,8 @@ class TripperTripListQueryControllerTest {
     @Test
     @DisplayName("미인증 사용자 요청 -> 인증 실패 401")
     @WithAnonymousUser
-    public void findTripperTripList_with_unauthorizedUser() throws Exception {
-        mockMvc.perform(get("/api/trips?tripper-id=1"))
+    public void findSingleSchedule_with_unauthorizedUser() throws Exception {
+        mockMvc.perform(get("/api/schedule-places/1"))
                 .andDo(print())
                 .andExpect(status().isUnauthorized())
                 .andExpect(jsonPath("$.errorCode").exists())
